+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### Bugs
 - Fixed TCP mux occasional freeze by adding write buffer to it.
 - Fixed stereo problem in chromium-based browsers, where it was only as mono by adding `stereo=1` to opus SDP to clients answer.
+- Fixed keysym mapping for unknown keycodes, which was causing some key combinations to not work on some keyboards.
 
 ## [n.eko v2.7](https://github.com/m1k1o/neko/releases/tag/v2.7)